dev.requirements.txt
pytest-django==3.5.0
requests-mock==1.6.0
black==19.3b0
pytest-mock==1.10.4
django-cors-headers==3.0.2
